(#1) Polish readability and autoload date formatter
