+++ composer.json
@@ -4,5 +4,8 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5"
+    },
+    "autoload": {
+        "files": ["src/DateFormatter.php"]
     }
 }
